Create Login page using html
To get started, we will first need to create a basic HTML file. In this file, we will include the main structure for our login form.


After creating the files just paste the following codes into your file. Make sure to save your HTML document with a .html extension, so that it can be properly viewed in a web browser.
This is the basic structure of our login form using HTML, and now we can move on to styling it using CSS.